feat(major): add all required entries for Tenant entity
As using NodeJS for backend we had to find a way to structurize the code on that  side. For that we used CLEAN architecture principles. This commit  embodies full realisation of one entity from database queries till routing REST api. We go from entities -> use cases (using db-source as composition) -> controllers -> routes

**!! Be aware this still needs testing and refactor. **
** Also keep attention to asyncF.ts useful utility and express-callback function